configure: Include time.h when checking for gmtime_r and localtime_r
These functions are available in time.h (conditional on posix thread
safe functions) on mingw.

Previously, these functions weren't detected by configure, and
libavutil/time_internal.h provided replacements, even if time.h
actually contained definitions of them.

These mingw inline functions are currently defined as
 "extern __inline __attribute__((__gnu_inline__))". In this case,
redefining a new static inline version of the same function with the
same name is tolerated.

+++ tests/fate-run.sh
@@ -8,7 +8,7 @@ base=$(dirname $0)
 base64=tests/base64
 
 test="${1#fate-}"
-target_samples=$2
+samples=$2
 target_exec=$3
 target_path=$4
 command=$5
@@ -25,7 +25,6 @@ cmp_unit=${15:-2}
 gen=${16:-no}
 hwaccel=${17:-none}
 report_type=${18:-standard}
-keep=${19:-0}
 
 outdir="tests/data/fate"
 outfile="${outdir}/${test}"
@@ -45,7 +44,7 @@ compare(){
 }
 
 do_tiny_psnr(){
-    psnr=$(tests/tiny_psnr "$1" "$2" $cmp_unit $cmp_shift 0) || return 1
+    psnr=$(tests/tiny_psnr "$1" "$2" $cmp_unit $cmp_shift 0)
     val=$(expr "$psnr" : ".*$3: *\([0-9.]*\)")
     size1=$(expr "$psnr" : '.*bytes: *\([0-9]*\)')
     size2=$(expr "$psnr" : '.*bytes:[ 0-9]*/ *\([0-9]*\)')
@@ -53,12 +52,6 @@ do_tiny_psnr(){
     size_cmp=$(compare $size1 $size2 $size_tolerance)
     if [ "$val_cmp" != 0 ] || [ "$size_cmp" != 0 ]; then
         echo "$psnr"
-        if [ "$val_cmp" != 0 ]; then
-            echo "$3: |$val - $cmp_target| >= $fuzz"
-        fi
-        if [ "$size_cmp" != 0 ]; then
-            echo "size: |$size1 - $size2| >= $size_tolerance"
-        fi
         return 1
     fi
 }
@@ -80,92 +73,42 @@ run(){
     $target_exec $target_path/"$@"
 }
 
-runecho(){
-    test "${V:-0}" -gt 0 && echo "$target_exec" $target_path/"$@" >&3
-    $target_exec $target_path/"$@" >&3
-}
-
 probefmt(){
-    run ffprobe${PROGSUF} -show_entries format=format_name -print_format default=nw=1:nk=1 -v 0 "$@"
+    run avprobe -show_format_entry format_name -v 0 "$@"
 }
 
-probetags(){
-    run ffprobe${PROGSUF} -show_entries format_tags -v 0 "$@"
+probestream(){
+    run avprobe -show_stream_entry "$1" -v 0 "$2"
 }
 
-runlocal(){
-    test "${V:-0}" -gt 0 && echo ${base}/"$@" ${base} >&3
-    ${base}/"$@" ${base}
-}
-
-probeframes(){
-    run ffprobe${PROGSUF} -show_frames -v 0 "$@"
-}
-
-probechapters(){
-    run ffprobe${PROGSUF} -show_chapters -v 0 "$@"
-}
-
-probegaplessinfo(){
-    filename="$1"
-    shift
-    run ffprobe${PROGSUF} -bitexact -select_streams a -show_entries format=start_time,duration:stream=index,start_pts,duration_ts -v 0 "$filename" "$@"
-    pktfile1="${outdir}/${test}.pkts"
-    framefile1="${outdir}/${test}.frames"
-    cleanfiles="$cleanfiles $pktfile1 $framefile1"
-    run ffprobe${PROGSUF} -bitexact -select_streams a -of compact -count_packets -show_entries packet=pts,dts,duration,flags:stream=nb_read_packets -v 0 "$filename" "$@" > "$pktfile1"
-    head -n 8 "$pktfile1"
-    tail -n 9 "$pktfile1"
-    run ffprobe${PROGSUF} -bitexact -select_streams a -of compact -count_frames -show_entries frame=pkt_pts,pkt_dts,best_effort_timestamp,pkt_duration,nb_samples:stream=nb_read_frames -v 0 "$filename" "$@" > "$framefile1"
-    head -n 8 "$framefile1"
-    tail -n 9 "$framefile1"
-}
-
-ffmpeg(){
+avconv(){
     dec_opts="-hwaccel $hwaccel -threads $threads -thread_type $thread_type"
-    ffmpeg_args="-nostdin -nostats -cpuflags $cpuflags"
+    avconv_args="-nostats -cpuflags $cpuflags"
     for arg in $@; do
-        [ x${arg} = x-i ] && ffmpeg_args="${ffmpeg_args} ${dec_opts}"
-        ffmpeg_args="${ffmpeg_args} ${arg}"
+        [ x${arg} = x-i ] && avconv_args="${avconv_args} ${dec_opts}"
+        avconv_args="${avconv_args} ${arg}"
     done
-    run ffmpeg${PROGSUF}${EXECSUF} ${ffmpeg_args}
+    run avconv ${avconv_args}
 }
 
 framecrc(){
-    ffmpeg "$@" -bitexact -f framecrc -
-}
-
-ffmetadata(){
-    ffmpeg "$@" -bitexact -f ffmetadata -
+    avconv "$@" -f framecrc -
 }
 
 framemd5(){
-    ffmpeg "$@" -bitexact -f framemd5 -
+    avconv "$@" -f framemd5 -
 }
 
 crc(){
-    ffmpeg "$@" -f crc -
-}
-
-md5pipe(){
-    ffmpeg "$@" md5:
+    avconv "$@" -f crc -
 }
 
 md5(){
-    encfile="${outdir}/${test}.out"
-    cleanfiles="$cleanfiles $encfile"
-    ffmpeg "$@" $encfile
-    do_md5sum $encfile | awk '{print $1}'
+    avconv "$@" md5:
 }
 
 pcm(){
-    ffmpeg "$@" -vn -f s16le -
-}
-
-fmtstdout(){
-    fmt=$1
-    shift 1
-    ffmpeg -bitexact "$@" -f $fmt -
+    avconv "$@" -vn -f s16le -
 }
 
 enc_dec_pcm(){
@@ -177,8 +120,8 @@ enc_dec_pcm(){
     encfile="${outdir}/${test}.${out_fmt}"
     cleanfiles=$encfile
     encfile=$(target_path ${encfile})
-    ffmpeg -i $src_file "$@" -f $out_fmt -y ${encfile} || return
-    ffmpeg -bitexact -i ${encfile} -c:a pcm_${pcm_fmt} -fflags +bitexact -f ${dec_fmt} -
+    avconv -i $src_file "$@" -f $out_fmt -y ${encfile} || return
+    avconv -f $out_fmt -i ${encfile} -c:a pcm_${pcm_fmt} -f ${dec_fmt} -
 }
 
 FLAGS="-flags +bitexact -sws_flags +accurate_rnd+bitexact -fflags +bitexact"
@@ -199,72 +142,35 @@ enc_dec(){
     tsrcfile=$(target_path $srcfile)
     tencfile=$(target_path $encfile)
     tdecfile=$(target_path $decfile)
-    ffmpeg -f $src_fmt $DEC_OPTS -i $tsrcfile $ENC_OPTS $enc_opt $FLAGS \
+    avconv -f $src_fmt $DEC_OPTS -i $tsrcfile $ENC_OPTS $enc_opt $FLAGS \
         -f $enc_fmt -y $tencfile || return
     do_md5sum $encfile
     echo $(wc -c $encfile)
-    ffmpeg $8 $DEC_OPTS -i $tencfile $ENC_OPTS $dec_opt $FLAGS \
+    avconv $DEC_OPTS -i $tencfile $ENC_OPTS $dec_opt $FLAGS \
         -f $dec_fmt -y $tdecfile || return
     do_md5sum $decfile
     tests/tiny_psnr $srcfile $decfile $cmp_unit $cmp_shift
 }
 
-transcode(){
-    src_fmt=$1
-    srcfile=$2
-    enc_fmt=$3
-    enc_opt=$4
-    final_decode=$5
-    encfile="${outdir}/${test}.${enc_fmt}"
-    test "$7" = -keep || cleanfiles="$cleanfiles $encfile"
-    tsrcfile=$(target_path $srcfile)
-    tencfile=$(target_path $encfile)
-    ffmpeg -f $src_fmt $DEC_OPTS -i $tsrcfile $ENC_OPTS $enc_opt $FLAGS \
-        -f $enc_fmt -y $tencfile || return
-    do_md5sum $encfile
-    echo $(wc -c $encfile)
-    ffmpeg $DEC_OPTS -i $encfile $ENC_OPTS $FLAGS $final_decode \
-        -f framecrc - || return
-}
-
-stream_remux(){
-    src_fmt=$1
-    srcfile=$2
-    enc_fmt=$3
-    stream_maps=$4
-    final_decode=$5
-    encfile="${outdir}/${test}.${enc_fmt}"
-    test "$7" = -keep || cleanfiles="$cleanfiles $encfile"
-    tsrcfile=$(target_path $srcfile)
-    tencfile=$(target_path $encfile)
-    ffmpeg -f $src_fmt -i $tsrcfile $stream_maps -codec copy $FLAGS \
-        -f $enc_fmt -y $tencfile || return
-    ffmpeg $DEC_OPTS -i $encfile $ENC_OPTS $FLAGS $final_decode \
-        -f framecrc - || return
-}
-
 # FIXME: There is a certain duplication between the avconv-related helper
 # functions above and below that should be refactored.
-ffmpeg2="$target_exec ${target_path}/ffmpeg${PROGSUF}${EXECSUF}"
+avconv2="$target_exec ${target_path}/avconv"
 raw_src="${target_path}/tests/vsynth1/%02d.pgm"
 pcm_src="${target_path}/tests/data/asynth1.sw"
 crcfile="tests/data/$test.lavf.crc"
 target_crcfile="${target_path}/$crcfile"
 
-[ "${V-0}" -gt 0 ] && echov=echov || echov=:
-
 echov(){
     echo "$@" >&3
 }
 
-AVCONV_OPTS="-nostdin -nostats -y -cpuflags $cpuflags"
-COMMON_OPTS="-flags +bitexact -idct simple -sws_flags +accurate_rnd+bitexact -fflags +bitexact"
-DEC_OPTS="$COMMON_OPTS -threads $threads"
-ENC_OPTS="$COMMON_OPTS -threads 1 -dct fastint"
+AVCONV_OPTS="-nostats -y -cpuflags $cpuflags -threads $threads"
+DEC_OPTS="-flags +bitexact -idct simple -sws_flags +accurate_rnd+bitexact -fflags +bitexact"
+ENC_OPTS="$DEC_OPTS -threads 1 -dct fastint"
 
 run_avconv(){
-    $echov $ffmpeg2 $AVCONV_OPTS $*
-    $ffmpeg2 $AVCONV_OPTS $*
+    $echov $avconv2 $AVCONV_OPTS $*
+    $avconv2 $AVCONV_OPTS $*
 }
 
 do_avconv(){
@@ -287,7 +193,7 @@ lavf_audio(){
     t="${test#lavf-}"
     outdir="tests/data/lavf"
     file=${outdir}/lavf.$t
-    do_avconv $file $DEC_OPTS $1 -ar 44100 -f s16le -i $pcm_src "$ENC_OPTS -metadata title=lavftest" -t 1 -qscale 10 $2
+    do_avconv $file $DEC_OPTS $1 -ar 44100 -f s16le -i $pcm_src $ENC_OPTS -t 1 -qscale 10 $2
     do_avconv_crc $file $DEC_OPTS $3 -i $target_path/$file
 }
 
@@ -295,30 +201,9 @@ lavf_container(){
     t="${test#lavf-}"
     outdir="tests/data/lavf"
     file=${outdir}/lavf.$t
-    do_avconv $file $DEC_OPTS -f image2 -c:v pgmyuv -i $raw_src $DEC_OPTS -ar 44100 -f s16le $1 -i $pcm_src "$ENC_OPTS -metadata title=lavftest" -b:a 64k -t 1 -qscale:v 10 $2
+    do_avconv $file $DEC_OPTS -f image2 -c:v pgmyuv -i $raw_src $DEC_OPTS -ar 44100 -f s16le $1 -i $pcm_src $ENC_OPTS -b:a 64k -t 1 -qscale:v 10 $2
     test $3 = "disable_crc" ||
         do_avconv_crc $file $DEC_OPTS -i $target_path/$file $3
-}
-
-lavf_container_attach() {          lavf_container "" "$1 -attach ${raw_src%/*}/00.pgm -metadata:s:t mimetype=image/x-portable-greymap"; }
-lavf_container_timecode_nodrop() { lavf_container "" "$1 -timecode 02:56:14:13"; }
-lavf_container_timecode_drop()   { lavf_container "" "$1 -timecode 02:56:14.13 -r 30000/1001"; }
-
-lavf_container_timecode()
-{
-    lavf_container_timecode_nodrop "$@"
-    lavf_container_timecode_drop "$@"
-    lavf_container "" "$1"
-}
-
-lavf_container_fate()
-{
-    t="${test#lavf-fate-}"
-    outdir="tests/data/lavf-fate"
-    file=${outdir}/lavf.$t
-    input="${target_samples}/$1"
-    do_avconv $file $DEC_OPTS $2 -i "$input" "$ENC_OPTS -metadata title=lavftest" -vcodec copy -acodec copy
-    do_avconv_crc $file $DEC_OPTS -i $target_path/$file $3
 }
 
 lavf_image(){
@@ -326,9 +211,9 @@ lavf_image(){
     outdir="tests/data/images/$t"
     mkdir -p "$outdir"
     file=${outdir}/%02d.$t
-    run_avconv $DEC_OPTS -f image2 -c:v pgmyuv -i $raw_src $1 "$ENC_OPTS -metadata title=lavftest" -frames 13 -y -qscale 10 $target_path/$file
+    run_avconv $DEC_OPTS -f image2 -c:v pgmyuv -i $raw_src $1 $ENC_OPTS $2 -frames 12 -y -qscale 10 $target_path/$file
     do_md5sum ${outdir}/02.$t
-    do_avconv_crc $file $DEC_OPTS $2 -i $target_path/$file $2
+    do_avconv_crc $file $DEC_OPTS $2 -i $target_path/$file
     echo $(wc -c ${outdir}/02.$t)
 }
 
@@ -337,7 +222,7 @@ lavf_image2pipe(){
     t="${t%pipe}"
     outdir="tests/data/lavf"
     file=${outdir}/${t}pipe.$t
-    do_avconv $file $DEC_OPTS -f image2 -c:v pgmyuv -i $raw_src -f image2pipe "$ENC_OPTS -metadata title=lavftest" -t 1 -qscale 10
+    do_avconv $file $DEC_OPTS -f image2 -c:v pgmyuv -i $raw_src -f image2pipe $ENC_OPTS -t 1 -qscale 10
     do_avconv_crc $file $DEC_OPTS -f image2pipe -i $target_path/$file
 }
 
@@ -345,17 +230,8 @@ lavf_video(){
     t="${test#lavf-}"
     outdir="tests/data/lavf"
     file=${outdir}/lavf.$t
-    do_avconv $file $DEC_OPTS -f image2 -c:v pgmyuv -i $raw_src "$ENC_OPTS -metadata title=lavftest" -t 1 -qscale 10 $1 $2
+    do_avconv $file $DEC_OPTS -f image2 -c:v pgmyuv -i $raw_src $ENC_OPTS -t 1 -qscale 10 $1
     do_avconv_crc $file $DEC_OPTS -i $target_path/$file $1
-}
-
-refcmp_metadata(){
-    refcmp=$1
-    pixfmt=$2
-    fuzz=${3:-0.001}
-    ffmpeg $FLAGS $ENC_OPTS \
-        -lavfi "testsrc2=size=300x200:rate=1:duration=5,format=${pixfmt},split[ref][tmp];[tmp]avgblur=4[enc];[enc][ref]${refcmp},metadata=print:file=-" \
-        -f null /dev/null | awk -v ref=${ref} -v fuzz=${fuzz} -f ${base}/refcmp-metadata.awk -
 }
 
 pixfmt_conversion(){
@@ -373,118 +249,38 @@ video_filter(){
     filters=$1
     shift
     label=${test#filter-}
-    raw_src="${target_path}/tests/vsynth1/%02d.pgm"
     printf '%-20s' $label
-    ffmpeg $DEC_OPTS -f image2 -vcodec pgmyuv -i $raw_src \
-        $FLAGS $ENC_OPTS -vf "$filters" -vcodec rawvideo -frames:v 5 $* -f nut md5:
+    avconv $DEC_OPTS -f image2 -c:v pgmyuv -i $raw_src \
+        $FLAGS $ENC_OPTS -vf "$filters" -c:v rawvideo -frames:v 5 $* -f nut md5:
 }
 
 pixfmts(){
     filter=${test#filter-pixfmts-}
-    filter=${filter%_*}
     filter_args=$1
-    prefilter_chain=$2
-    nframes=${3:-1}
 
     showfiltfmts="$target_exec $target_path/libavfilter/tests/filtfmts"
-    scale_exclude_fmts=${outfile}_scale_exclude_fmts
-    scale_in_fmts=${outfile}_scale_in_fmts
-    scale_out_fmts=${outfile}_scale_out_fmts
-    in_fmts=${outfile}_in_fmts
+    exclude_fmts=${outfile}${filter}_exclude_fmts
+    out_fmts=${outfile}${filter}_out_fmts
 
     # exclude pixel formats which are not supported as input
-    $showfiltfmts scale | awk -F '[ \r]' '/^INPUT/{ fmt=substr($3, 5); print fmt }' | sort >$scale_in_fmts
-    $showfiltfmts scale | awk -F '[ \r]' '/^OUTPUT/{ fmt=substr($3, 5); print fmt }' | sort >$scale_out_fmts
-    comm -12 $scale_in_fmts $scale_out_fmts >$scale_exclude_fmts
+    avconv -pix_fmts list 2>/dev/null | awk 'NR > 8 && /^\..\./ { print $2 }' | sort >$exclude_fmts
+    $showfiltfmts scale | awk -F '[ \r]' '/^OUTPUT/{ print $3 }' | sort | comm -23 - $exclude_fmts >$out_fmts
 
-    $showfiltfmts $filter | awk -F '[ \r]' '/^INPUT/{ fmt=substr($3, 5); print fmt }' | sort >$in_fmts
-    pix_fmts=$(comm -12 $scale_exclude_fmts $in_fmts)
+    pix_fmts=$($showfiltfmts $filter | awk -F '[ \r]' '/^INPUT/{ print $3 }' | sort | comm -12 - $out_fmts)
 
     outertest=$test
     for pix_fmt in $pix_fmts; do
         test=$pix_fmt
-        video_filter "${prefilter_chain}format=$pix_fmt,$filter=$filter_args" -pix_fmt $pix_fmt -frames:v $nframes
+        video_filter "format=$pix_fmt,$filter=$filter_args" -pix_fmt $pix_fmt -frames:v 1
     done
 
-    rm $in_fmts $scale_in_fmts $scale_out_fmts $scale_exclude_fmts
+    rm $exclude_fmts $out_fmts
     test=$outertest
-}
-
-gapless(){
-    sample=$(target_path $1)
-    extra_args=$2
-
-    decfile1="${outdir}/${test}.out-1"
-    decfile2="${outdir}/${test}.out-2"
-    decfile3="${outdir}/${test}.out-3"
-    cleanfiles="$cleanfiles $decfile1 $decfile2 $decfile3"
-
-    # test packet data
-    ffmpeg $extra_args -i "$sample" -bitexact -c:a copy -f framecrc -y $decfile1
-    do_md5sum $decfile1
-    # test decoded (and cut) data
-    ffmpeg $extra_args -i "$sample" -bitexact -f wav md5:
-    # the same as above again, with seeking to the start
-    ffmpeg $extra_args -ss 0 -seek_timestamp 1 -i "$sample" -bitexact -c:a copy -f framecrc -y $decfile2
-    do_md5sum $decfile2
-    ffmpeg $extra_args -ss 0 -seek_timestamp 1 -i "$sample" -bitexact -f wav md5:
-    # test packet data, with seeking to a specific position
-    ffmpeg $extra_args -ss 5 -seek_timestamp 1 -i "$sample" -bitexact -c:a copy -f framecrc -y $decfile3
-    do_md5sum $decfile3
-}
-
-gaplessenc(){
-    sample=$(target_path $1)
-    format=$2
-    codec=$3
-
-    file1="${outdir}/${test}.out-1"
-    cleanfiles="$cleanfiles $file1"
-
-    # test data after reencoding
-    ffmpeg -i "$sample" -bitexact -map 0:a -c:a $codec -f $format -y "$file1"
-    probegaplessinfo "$file1"
-}
-
-audio_match(){
-    sample=$(target_path $1)
-    trefile=$(target_path $2)
-    extra_args=$3
-
-    decfile="${outdir}/${test}.wav"
-    cleanfiles="$cleanfiles $decfile"
-
-    ffmpeg -i "$sample" -bitexact $extra_args -y $decfile
-    tests/audiomatch $decfile $trefile
-}
-
-concat(){
-    template=$1
-    sample=$2
-    mode=$3
-    extra_args=$4
-
-    concatfile="${outdir}/${test}.ffconcat"
-    packetfile="${outdir}/${test}.ffprobe"
-    cleanfiles="$concatfile $packetfile"
-
-    awk "{gsub(/%SRCFILE%/, \"$sample\"); print}" $template > $concatfile
-
-    if [ "$mode" = "md5" ]; then
-        run ffprobe${PROGSUF} -bitexact -show_streams -show_packets -v 0 -fflags keepside -safe 0 $extra_args $concatfile | tr -d '\r' > $packetfile
-        do_md5sum $packetfile
-    else
-        run ffprobe${PROGSUF} -bitexact -show_streams -show_packets -v 0 -of compact=p=0:nk=1 -fflags keepside -safe 0 $extra_args $concatfile
-    fi
 }
 
 null(){
     :
 }
-
-# Disable globbing: command arguments may contain globbing characters and
-# must be kept verbatim
-set -f
 
 exec 3>&2
 eval $command >"$outfile" 2>$errfile
@@ -495,14 +291,12 @@ if [ $err -gt 128 ]; then
     test "${sig}" = "${sig%[!A-Za-z]*}" || unset sig
 fi
 
-if test -e "$ref" || test $cmp = "oneline" || test $cmp = "null" || test $cmp = "grep" ; then
+if test -e "$ref" || test $cmp = "oneline" || test $cmp = "null" ; then
     case $cmp in
         diff)   diff -u -b "$ref" "$outfile"            >$cmpfile ;;
-        rawdiff)diff -u    "$ref" "$outfile"            >$cmpfile ;;
         oneoff) oneoff     "$ref" "$outfile"            >$cmpfile ;;
         stddev) stddev     "$ref" "$outfile"            >$cmpfile ;;
         oneline)oneline    "$ref" "$outfile"            >$cmpfile ;;
-        grep)   grep       "$ref" "$errfile"            >$cmpfile ;;
         null)   cat               "$outfile"            >$cmpfile ;;
     esac
     cmperr=$?
@@ -531,14 +325,5 @@ if test $err != 0 && test $gen != "no" ; then
     err=$?
 fi
 
-if test $err = 0; then
-    if test $keep = 0; then
-        rm -f $outfile $errfile $cmpfile $cleanfiles
-    fi
-elif test $gen = "no"; then
-    echo "Test $test failed. Look at $errfile for details."
-    test "${V:-0}" -gt 0 && cat $errfile
-else
-    echo "Updating reference failed, possibly no output file was generated."
-fi
+test $err = 0 && rm -f $outfile $errfile $cmpfile $cleanfiles
 exit $err
